- Refacor task and project into projectDom and taskDom
